docker_images/api/crawler_management_controller/requirements.txt
requests==2.25.1
requests-oauthlib==1.3.1
uvicorn==0.19.0
fastapi==0.87.0
classy-fastapi==0.2.12
fastapi-utils==0.2.1
starlette==0.21.0
pyfunctional==1.4.3
pydantic==1.10.2
python-jose==3.3.0